requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
google-generativeai==0.3.2
python-dotenv==1.0.0
markdown==3.5.1
werkzeug==2.3.7
